+++ huaweicloud-dis-kafka-adapter-0.11/src/test/java/com/huaweicloud/dis/adapter/kafka/DisKafkaProducerDemo.java
@@ -20,12 +20,12 @@ import java.nio.ByteBuffer;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
+import com.huaweicloud.dis.adapter.kafka.producer.DISKafkaProducer;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 
 import com.huaweicloud.dis.DISConfig;
-import com.huaweicloud.dis.adapter.kafka.producer.DISKafkaProducer;
 
 public class DisKafkaProducerDemo
 {
@@ -34,7 +34,7 @@ public class DisKafkaProducerDemo
         DISConfig disConfig = new DISConfig();
         disConfig.set("IS_DEFAULT_TRUSTED_JKS_ENABLED", "false");
         
-        Producer<String, byte[]> disKafkaProducer = new DISKafkaProducer<String, byte[]>(disConfig);
+        Producer<String, byte[]> disKafkaProducer = new DISKafkaProducer<>(disConfig);
         
         // 配置流名称，根据实际情况修改为与开通的DIS通道一致的“通道名称”
         String streamName = "zj-ttt";