+++ huaweicloud-dis-kafka-adapter-0.11/src/test/java/com/huaweicloud/dis/adapter/kafka/consumer/consumerKVTest.java
@@ -34,8 +34,8 @@ public class consumerKVTest {
     public void KVTest0()
     {
         DISConfig disConfig = new DISConfig();
-        disConfig.set("key.deserializer","org.apache.common.common.serialization.StringDeserializer");
-        disConfig.set("value.deserializer","org.apache.common.common.serialization.StringDeserializer");
+        disConfig.set("key.deserializer","org.apache.kafka.common.serialization.StringDeserializer");
+        disConfig.set("value.deserializer","org.apache.kafka.common.serialization.StringDeserializer");
         DISKafkaConsumer<String,String> disKafkaConsumer = new DISKafkaConsumer<String, String>(disConfig);
     }
 
@@ -64,16 +64,16 @@ public class consumerKVTest {
     public void KVTest3()
     {
         DISConfig disConfig = new DISConfig();
-        disConfig.set("key.deserializer","org.apache.common.common.serialization.LongDeserializer");
-        disConfig.set("value.deserializer","org.apache.common.common.serialization.StringDeserializer");
+        disConfig.set("key.deserializer","org.apache.kafka.common.serialization.LongDeserializer");
+        disConfig.set("value.deserializer","org.apache.kafka.common.serialization.StringDeserializer");
         DISKafkaConsumer<String,String> disKafkaConsumer = new DISKafkaConsumer<String, String>(disConfig);
     }
     @Test
     public void KVTest4()
     {
         DISConfig disConfig = new DISConfig();
-        disConfig.set("key.deserializer","org.apache.common.common.serialization.StringDeserializer");
-        disConfig.set("value.deserializer","org.apache.common.common.serialization.IntegerDeserializer");
+        disConfig.set("key.deserializer","org.apache.kafka.common.serialization.StringDeserializer");
+        disConfig.set("value.deserializer","org.apache.kafka.common.serialization.IntegerDeserializer");
         DISKafkaConsumer<String,String> disKafkaConsumer = new DISKafkaConsumer<String, String>(disConfig);
     }