+++ huaweicloud-dis-kafka-adapter-common/src/main/java/com/huaweicloud/dis/adapter/common/model/PartitionIterator.java
@@ -1,0 +1,4 @@
+package com.huaweicloud.dis.adapter.common.model;
+
+public class PartitionIterator {
+}