+++ spring-cloud-streams-kafka-dis-adapter-demo/src/main/java/demo/GreetingsService.java
@@ -1,0 +1,30 @@
+package demo;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.MessageHeaders;
+import org.springframework.messaging.support.MessageBuilder;
+import org.springframework.stereotype.Service;
+import org.springframework.util.MimeTypeUtils;
+
+@Service
+public class GreetingsService {
+    private static Logger log = LoggerFactory.getLogger(GreetingsService.class);
+    
+    private final GreetingsStreams greetingsStreams;
+
+    public GreetingsService(GreetingsStreams greetingsStreams) {
+        this.greetingsStreams = greetingsStreams;
+    }
+
+    public void sendGreeting(final Greetings greetings) {
+        log.info("Sending greetings {}", greetings);
+
+        MessageChannel messageChannel = greetingsStreams.outboundGreetings();
+        messageChannel.send(MessageBuilder
+                .withPayload(greetings)
+                .setHeader(MessageHeaders.CONTENT_TYPE, MimeTypeUtils.APPLICATION_JSON)
+                .build());
+    }
+}