+++ spring-cloud-streams-kafka-dis-adapter-demo/src/main/java/demo/StreamsConfig.java
@@ -1,0 +1,7 @@
+package demo;
+
+import org.springframework.cloud.stream.annotation.EnableBinding;
+
+@EnableBinding(GreetingsStreams.class)
+public class StreamsConfig {
+}