+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 - Features
   * add kafka clients 1.1.0 supports
   * add kafka clients 0.11 supports
-  * add common adapter
   
+# 1.1.2
+- Features
+  * add dis spring kafka demo
+
+# 1.1.3
+- Features
+  * add common adapter